# openapi.yaml
openapi: 3.1.0
info:
  title: Human Design API
  version: 3.4.1
paths:
...
  /panel/login:
    get:
      tags:
      - panel
      summary: Login Page
      operationId: login_page_panel_login_get
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
    post:
      tags:
      - panel
      summary: Login Submit
      operationId: login_submit_panel_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_submit_panel_login_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /panel/dashboard:
    get:
      tags:
      - panel
      summary: Dashboard
      operationId: dashboard_panel_dashboard_get
      parameters:
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          default: 30
          title: Days
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /panel/api/sites/{site_id}:
    put:
      tags:
      - panel
      summary: Api Update Site
      operationId: api_update_site_panel_api_sites__site_id__put
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: integer
          title: Site Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - panel
      summary: Api Delete Site
      operationId: api_delete_site_panel_api_sites__site_id__delete
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: integer
          title: Site Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /panel/api/logs:
    get:
      tags:
      - panel
      summary: Api Get Logs
      operationId: api_get_logs_panel_api_logs_get
      parameters:
      - name: site_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Site Id
      - name: endpoint
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Endpoint
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_login_submit_panel_login_post:
      properties:
        token:
          type: string
          title: Token
      type: object
      required:
      - token
      title: Body_login_submit_panel_login_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer